# docs/swagger.yaml
basePath: /api/v1
info:
  contact: {}
  description: REST API for SL651 Device Management Platform
  title: SL651 Platform API
  version: "1.0"
paths:
  /devices:
    get:
      description: Get a list of all registered devices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      description: Register a new device
      parameters:
      - description: Device info
        in: body
        name: device
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create device
      tags:
      - devices
  /devices/{id}:
    delete:
      description: Remove a device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete device
      tags:
      - devices
    get:
      description: Get detailed information of a device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get device
      tags:
      - devices
    put:
      consumes:
      - application/json
      description: Update device information
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Device updates
        in: body
        name: device
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update device
      tags:
      - devices
  /devices/{id}/data:
    get:
      description: Retrieve historical data for a device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get device data
      tags:
      - devices
  /devices/{id}/data/latest:
    get:
      description: Retrieve the most recent data point for a device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get latest device data
      tags:
      - devices
  /health:
    get:
      description: Check server health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Health Check
      tags:
      - health
  /statistics/devices:
    get:
      description: Get platform-wide device statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get device statistics
      tags:
      - statistics
  /statistics/forward:
    get:
      description: Get forwarding service statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get forward statistics
      tags:
      - statistics
swagger: "2.0"